pyproject.toml
[tool.poetry]
name = "async_api_sprint_1"
version = "0.1.0"
description = ""
authors = ["Sergey Klimov", "Maxim Rukhlov ", "Alex Erdyakov <sa@prg.re>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
aioredis = "^2.0.0"
backoff = "^1.11.1"
fastapi = "^0.68.0"
orjson = "^3.6.1"
fastapi-cache2 = "^0.1.6"
fastapi-pagination = "^0.8.3"
uvicorn = "^0.14.0"
elasticsearch-async = "^6.2.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2.4"
pytest-asyncio = "^0.15.1"
aiohttp = "^3.7.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
